move version from config tab to location tab, rename config networking and location config
